+++ pkg/codegen/go/gen_program_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
+	"github.com/pulumi/pulumi/pkg/v2/codegen"
 	"github.com/pulumi/pulumi/pkg/v2/codegen/hcl2"
 	"github.com/pulumi/pulumi/pkg/v2/codegen/hcl2/model/format"
 	"github.com/pulumi/pulumi/pkg/v2/codegen/hcl2/syntax"
@@ -28,6 +29,7 @@ func TestGenProgram(t *testing.T) {
 		}
 		// TODO: include all test files
 		if filepath.Base(f.Name()) != "aws-s3-logging.pp" &&
+			filepath.Base(f.Name()) != "aws-s3-folder.pp" &&
 			filepath.Base(f.Name()) != "aws-fargate.pp" {
 			continue
 		}
@@ -63,7 +65,7 @@ func TestGenProgram(t *testing.T) {
 			files, diags, err := GenerateProgram(program)
 			assert.NoError(t, err)
 			if diags.HasErrors() {
-				t.Fatalf("failed to bind program: %v", diags)
+				t.Fatalf("failed to generate program: %v", diags)
 			}
 			assert.Equal(t, string(expected), string(files["main.go"]))
 		})
@@ -116,9 +118,11 @@ func newTestGenerator(t *testing.T, testFile string) *generator {
 		}
 
 		g := &generator{
-			program:            program,
-			jsonTempSpiller:    &jsonSpiller{},
-			ternaryTempSpiller: &tempSpiller{},
+			program:             program,
+			jsonTempSpiller:     &jsonSpiller{},
+			ternaryTempSpiller:  &tempSpiller{},
+			readDirTempSpiller:  &readDirSpiller{},
+			scopeTraversalRoots: codegen.NewStringSet(),
 		}
 		g.Formatter = format.NewFormatter(g)
 		return g